Two improvements to margo controller
Add watch deployment manifests by label;
Remove the setNamespace function. This is problematic because k-d-p does not apply `WithObjectTransform` in `reconcileExists` when it tries to find the existing manfiests. That means if the manifests from local-dir does not have namespace, k-d-p won't find them from the cluster. So I added namespace to the manifests directly to avoid confusion

Change-Id: If31f8c21d7f1c9e32dbbced9bbf43d7d077d3e1e