fix: specify default fields to avoid Server-Side-Apply no-op modification
NOTE: The manifest changes is from a kustomize patch but we do not want to put the kustomize config to the channel dir. We should have a CI (e.g. a make rule that run kustomize in another location) to update the ArgoCD manifests.

When using ApplysetApplier, the Moss watches 2 NetworkPolicy objects and 1 Deployment object being "MODIFIED" (even if the manifests are not changed), and triggers repeated reconciliation.
This is caused when the NetworkPolicy `procotol` (should be optional, default to TCP) and the Deployment `spec.containers.env.valueFrom.fieldRef.apiVersion` (should be optional, default to v1) are not specified.

b/277647949

Change-Id: Ieb28437a692bd3c6599ee361211421b3d85d3f7b